backend/requirements.txt
# Dev environment
pip
autopep8

# App
flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
Flask-WTF
email-validator
python-dotenv
psycopg2-binary
